- Removed footer - CSS updates - Separated weight display component in new file
