+++ src/App.css
@@ -6,7 +6,7 @@
     display: flex;
     min-height: 100vh;
     flex-direction: column;
-    justify-content: space-between;
+    justify-content: center;
   }
   
   .App-logo {
