+++ src/components/WeightInput/WeightInput.css
@@ -12,11 +12,47 @@
     align-items: center;
 }
 
-.weightDisplay * {
+.weightDisplay {
+    height: 150px;
+    /* max-width: 100%; */
+    /* max-height: 100%; */
+}
+
+.plateImg {
     max-width: 100%;
     max-height: 100%;
 }
 
+.imgBox {
+    border: 5px dotted lightgray;
+    width: 50%;
+}
+
+.weightDisplayNumber {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    
+    text-shadow:
+        -1px -1px 0 black,
+        1px -1px 0 black,
+        -1px 1px 0 black,
+        1px 1px 0 black;
+    color: white;
+    font-weight: 900;
+    display:inline-block;
+    font-size: 2em;
+    /* background-color: lightslategray;
+    padding: .5rem 1.5rem .5rem 1.5rem;
+    border-radius: 9999px; */
+}
+
 td {
-    padding: 2.5vh;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding: 5px;
+    position: relative;
+    height: 100%;
 }
